requirements.txt
azure-cognitiveservices-speech
python-dotenv
azure-ai-textanalytics
fastapi
uvicorn
requests
